# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HotelInfo {
  name: string;
  location: string;
  checkInTime: string;
  checkOutTime: string;
  amenities: string[];
  nearbyAttractions: string[];
  policies: string[];
}

export interface Room {
  id: number;
  name: string;
  type: string;
  pricePerNight: number;
  description: string;
  amenities: string[];
  maxGuests: number;
  available: boolean;
  /** @nullable */
  imageUrl?: string | null;
}

export interface BookingInput {
  guestName: string;
  phone: string;
  email: string;
  checkInDate: string;
  checkOutDate: string;
  numberOfGuests: number;
  roomType: string;
  roomId: number;
  /** @nullable */
  specialRequests?: string | null;
}

export interface Booking {
  id: number;
  guestName: string;
  phone: string;
  email: string;
  checkInDate: string;
  checkOutDate: string;
  numberOfGuests: number;
  roomType: string;
  roomId: number;
  /** @nullable */
  specialRequests?: string | null;
  status: string;
  totalPrice: number;
  confirmationCode: string;
  createdAt: string;
}

export interface BookingStats {
  totalBookings: number;
  confirmedBookings: number;
  totalRevenue: number;
  popularRoomType: string;
}

export interface ApiError {
  error: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiError {
  error: string;
}