# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  guestName: string;
  phone: string;
  email: string;
  checkInDate: string;
  checkOutDate: string;
  numberOfGuests: number;
  roomType: string;
  roomId: number;
  /** @nullable */
  specialRequests?: string | null;
  status: string;
  totalPrice: number;
  confirmationCode: string;
  createdAt: Date;
}