# lib/api-zod/src/generated/types/hotelInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HotelInfo {
  name: string;
  location: string;
  checkInTime: string;
  checkOutTime: string;
  amenities: string[];
  nearbyAttractions: string[];
  policies: string[];
}